udpate

+++ Chess3D/Assets/Scripts/UI/Gameplay/UIGameplaySlider.cs
@@ -95,7 +95,8 @@ public class UIGameplaySlider : MonoBehaviour
 
     public void ChangeValueFromTurn(int turn, int maxTurn)
     {
-        slider.value = turn / maxTurn;
+        
+        slider.value = (float)turn / maxTurn;
         
     }
 }
